+++ src/LionWeb.Core/Core/Migration/MigrationExceptions.cs
@@ -57,13 +57,12 @@ public class InvalidRootNodesException(IMigration migration, string message)
 public class InvalidMigrationResultException(string message)
     : MigrationExceptionBase(message);
 
-/// Not exactly one language could be mapped to the same <paramref name="key"/>.
-public class AmbiguousLanguageKeyMapping(string key, IEnumerable<DynamicLanguage> languages)
-    : MigrationExceptionBase($"Not exactly one mapped language for key {key}: {string.Join(", ", languages.Select(l => $"({l.Key}, {l.Version})"))}");
-
 /// Lookup of a <see cref="IKeyed"/> language element failed. 
 public class UnknownLookupException : MigrationExceptionBase
 {
+    /// No equivalent can be found for <paramref name="key"/>.
+    public UnknownLookupException(string key, LanguageIdentity languageIdentity) : base($"No lookup found for {languageIdentity}, key {key}") { }
+    
     /// No equivalent can be found for <paramref name="key"/>.
     public UnknownLookupException(string key) : base($"No lookup found for key {key}") { }