+++ src/LionWeb.Core/Core/Migration/MigrationExtensions.cs
@@ -20,6 +20,7 @@ namespace LionWeb.Core.Migration;
 using M1;
 using M2;
 using M3;
+using System.Diagnostics.CodeAnalysis;
 
 /// Extensions useful to work with <see cref="LenientNode"/>s during Migration.
 public static class MigrationExtensions
@@ -32,19 +33,6 @@ public static class MigrationExtensions
     public static IEnumerable<LenientNode> Descendants(this LenientNode node) =>
         M1Extensions.Descendants(node, true, true);
 
-    /// All <see cref="Descendants(System.Collections.Generic.List{LionWeb.Core.LenientNode})">descendants</see> of
-    /// <paramref name="nodes"/> that are instances of <paramref name="classifier"/>.
-    [Obsolete]
-    public static IEnumerable<LenientNode>
-        AllInstancesOf(this List<LenientNode> nodes, ClassifierIdentity classifier) =>
-        nodes
-            .Descendants()
-            .Where(n =>
-            {
-                var classifierIdentity = ClassifierIdentity.FromClassifier(n.GetClassifier());
-                return classifierIdentity == classifier;
-            });
-
     /// Collects all languages used in <paramref name="nodes"/> via their classifier + generalizations, features, types,
     /// and instance values (for <see cref="EnumerationLiteral"/>).  
     public static IEnumerable<DynamicLanguage> CollectUsedLanguages(List<LenientNode> nodes)
@@ -53,7 +41,7 @@ public static class MigrationExtensions
 
         foreach (var node in nodes.Descendants())
             CollectUsedLanguages(node.GetClassifier(), result);
-        
+
         return result
             .Where(l => l is not IBuiltInsLanguage and not ILionCoreLanguage)
             .Cast<DynamicLanguage>();
@@ -67,7 +55,7 @@ public static class MigrationExtensions
         alreadyCollected.Add(classifier.GetLanguage());
         foreach (var feature in classifier.Features)
             CollectUsedLanguages(feature, alreadyCollected);
-        
+
         switch (classifier)
         {
             case Annotation a:
@@ -76,13 +64,13 @@ public static class MigrationExtensions
                 foreach (var iface in a.Implements)
                     CollectUsedLanguages(iface, alreadyCollected);
                 break;
-            
+
             case Concept c:
                 CollectUsedLanguages(c.Extends, alreadyCollected);
                 foreach (var iface in c.Implements)
                     CollectUsedLanguages(iface, alreadyCollected);
                 break;
-            
+
             case Interface i:
                 foreach (var iface in i.Extends)
                     CollectUsedLanguages(iface, alreadyCollected);
@@ -94,9 +82,9 @@ public static class MigrationExtensions
     {
         if (feature == null)
             return;
-        
+
         alreadyCollected.Add(feature.GetLanguage());
-        
+
         switch (feature)
         {
             case Property p:
@@ -112,7 +100,7 @@ public static class MigrationExtensions
     {
         if (datatype == null)
             return;
-        
+
         alreadyCollected.Add(datatype.GetLanguage());
 
         if (datatype is not StructuredDataType s)
@@ -121,50 +109,20 @@ public static class MigrationExtensions
         foreach (var field in s.Fields)
             CollectUsedLanguages(field.Type, alreadyCollected);
     }
-    //
-    // private static void CollectUsedLanguages(LenientNode? node, HashSet<Language> alreadyCollected)
-    // {
-    //     if (node == null)
-    //         return;
-    //     
-    //     foreach (var property in node.CollectAllSetFeatures().OfType<Property>())
-    //     {
-    //         if (node.TryGet(property, out var value))
-    //             CollectUsedLanguages(property.Type, value, alreadyCollected);
-    //         
-    //     }
-    // }
-    //
-    // private static void CollectUsedLanguages(Datatype datatype, object? value, HashSet<Language> alreadyCollected)
-    // {
-    //     switch (datatype)
-    //     {
-    //         case Enumeration e:
-    //             x = CollectUsedLanguages(e);
-    //             break;
-    //         case StructuredDataType:
-    //             x = value is IStructuredDataTypeInstance inst
-    //                 ? inst.CollectAllSetFields().SelectMany(f => CollectUsedLanguages(f.Type, inst.Get(f)))
-    //                 : [];
-    //             break;
-    //         default:
-    //             x = [];
-    //             break;
-    //     }
-    // }
 
-    internal static T Lookup<T>(T keyed, Language language) where T : IKeyed
+    internal static bool TryLookup<T>(string key, Language language, [NotNullWhen(true)] out T? result) where T : IKeyed
     {
         var mapped = M1Extensions
             .Descendants<IKeyed>(language, true)
-            .FirstOrDefault(k => k.Key == keyed.Key);
+            .FirstOrDefault(k => k.Key == key);
 
-        if (mapped is T result)
-            return result;
+        if (mapped is T r)
+        {
+            result = r;
+            return true;
+        }
 
-        if (mapped != null)
-            throw new UnknownLookupException($"{mapped.Key}, as it's not a {typeof(T)}: {keyed}");
-        
-        throw new UnknownLookupException(keyed);
+        result = default;
+        return false;
     }
 }