+++ test/LionWeb.Core.Test/Migration/DynamicLanguageClonerTests.cs
@@ -63,7 +63,7 @@ public class DynamicLanguageClonerTests
         var cloner = new DynamicLanguageCloner(LionWebVersions.Current);
         var dynamicLanguages = cloner.Clone([inputLang]);
 
-        Assert.IsInstanceOfType<MigrationFactory>(dynamicLanguages.Values.First().GetFactory());
+        Assert.IsNotInstanceOfType<MigrationFactory>(dynamicLanguages.Values.First().GetFactory());
     }
 
     [TestMethod]
@@ -101,6 +101,28 @@ public class DynamicLanguageClonerTests
         var cloner = new DynamicLanguageCloner(LionWebVersions.Current);
 
         Assert.ThrowsException<UnknownLookupException>(() => cloner.Clone([inputLangA]));
+    }
+
+    [TestMethod]
+    public void Language_ExternalDependsOn_Listed()
+    {
+        var inputLangA = new DynamicLanguage("langA", LionWebVersions.Current)
+        {
+            Key = "key-langA", Name = "langA", Version = "version"
+        };
+        var inputLangB = new DynamicLanguage("langB", LionWebVersions.Current)
+        {
+            Key = "key-langB", Name = "langB", Version = "version"
+        };
+        inputLangA.AddDependsOn([inputLangB]);
+
+        var cloner = new DynamicLanguageCloner(LionWebVersions.Current, [inputLangB]);
+
+        var dynamicLanguages = cloner.Clone([inputLangA]);
+
+        AssertEqual([inputLangA], dynamicLanguages);
+        Assert.AreEqual(1, dynamicLanguages.Values.Count);
+        Assert.AreSame(inputLangB, dynamicLanguages.Values.First().DependsOn[0]);
     }
 
     [TestMethod]